backend/requirements.txt
asgiref==3.7.1
click==8.1.3
channels==4.0.0
channels-redis==4.1.0
Django==3.2
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
mypy-extensions==1.0.0
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
celery==5.3.6
daphne==4.0.0
pyaes==1.6.1
pydantic==1.10.7
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
TgCrypto==1.2.5
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
choicelib==0.1.5
exceptiongroup==1.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.3
vbml==1.1.post1
vkbottle==4.3.12
vkbottle-types==5.131.146.16
watchfiles==0.18.1
yarl==1.9.2
deeppavlov==1.1.1